+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     *   Add option to switch to the HLS video stream when playing a downloaded episode
         ([#5620](https://github.com/Automattic/pocket-casts-android/pull/5620))
 *   Bug Fixes
+    *   Fix Bluetooth play controls resuming slowly and playback repeatedly pausing on wireless Android Auto
+        ([#5645](https://github.com/Automattic/pocket-casts-android/pull/5645))
     *   Allow copying description and show notes links with a long press
         ([#5628](https://github.com/Automattic/pocket-casts-android/pull/5628))